Rename header files to avoid conflicts with fitacf3.0 header files with same names
