+++ codebase/superdarn/src.lib/tk/lmfit_v2.0/include/lmfit_determinations.h
@@ -32,7 +32,7 @@ Modifications:
 #ifndef _ACFDETERMINATIONS_H
 #define _ACFDETERMINATIONS_H
 
-#include "fit_structures.h"
+#include "lmfit_structures.h"
 #include "fitdata.h"
 #include "fitblk.h"
 #include "llist.h"
