+++ codebase/superdarn/src.lib/tk/lmfit_v2.0/include/lmfit_fitting.h
@@ -33,7 +33,7 @@ Modifications:
 #define _ACFFIT_H
 
 #include "llist.h"
-#include "fit_structures.h"
+#include "lmfit_structures.h"
 
 typedef enum{
 	ACF,
