+++ codebase/superdarn/src.lib/tk/lmfit_v2.0/src/lmfit_determinations.c
@@ -27,7 +27,7 @@ Modifications:
 
 */
 
-#include "determinations.h"
+#include "lmfit_determinations.h"
 #include <math.h>
 #include <stdio.h>
 #include <stdlib.h>
